added try-catch block for sub-refresh-invoice-task to handle errors and show them on ui, and it helps to unstuck invoices in updating status
